refactor: user, auth modules 분리
