+++ src/frontend/.env
@@ -1,4 +1,4 @@
 NEXT_PUBLIC_BASE_URL_CHARGE = http://3.35.0.63:8080
 NEXT_PUBLIC_BASE_URL_MEMBERSHIP = http://localhost:8080
 NEXT_PUBLIC_BASE_URL_STORE = http://54.180.117.120:8100
-NEXT_PUBLIC_BASE_URL_AUTH = http://localhost:8001
+NEXT_PUBLIC_BASE_URL_AUTH = http://localhost:8101
