+++ src/frontend/api/user/type.ts
@@ -11,7 +11,6 @@ export interface IGetWithFriendReqType {
 export interface IResStoreType {
   data: any; //API request를 통한 서버응답이 아닌 store만 거치는 경우를 위한 return 타입
 }
-
 export interface ISaveUserInfoReqtype {
   email: string;
   country: string;
