+++ src/frontend/modules/auth/index.ts
@@ -1,0 +1,4 @@
+export { default } from './reducer';
+export * from './actions';
+export * from './types';
+export * from './saga';
