+++ src/frontend/modules/auth/initalData.ts
@@ -1,0 +1,1 @@
+export const initialSingup = {};
