+++ src/frontend/modules/auth/types.ts
@@ -1,0 +1,10 @@
+import { ActionType } from 'typesafe-actions';
+
+import { AsyncState } from 'modules/utils/reducerUtils';
+import * as actions from './actions';
+
+export interface IAuthState {
+  signup: AsyncState<{}, Error>;
+}
+
+export type authAction = ActionType<typeof actions>;
