+++ src/frontend/modules/user/reducer.ts
@@ -1,35 +1,15 @@
 import { createReducer } from 'typesafe-actions';
 
 import { asyncState } from 'modules/utils/reducerUtils';
-import { initialSingup, initalUserInfo } from 'modules/user/initalData';
+import { initalUserInfo } from 'modules/user/initalData';
 import { IUserState } from './types';
-import {
-  DO_SIGNUP,
-  DO_SIGNUP_SUCCESS,
-  DO_SIGNUP_FAIL,
-  SAVE_USERINFO,
-  SAVE_USERINFO_SUCCESS,
-  SAVE_USERINFO_FAIL,
-} from './actions';
+import { SAVE_USERINFO, SAVE_USERINFO_SUCCESS, SAVE_USERINFO_FAIL } from './actions';
 
 const initialState: IUserState = {
-  signup: asyncState.initial(initialSingup),
   userInfo: asyncState.initial(initalUserInfo),
 };
 
 const reducer = createReducer<IUserState>(initialState, {
-  [DO_SIGNUP]: (state, action) => ({
-    ...state,
-    signup: asyncState.load(initialSingup),
-  }),
-  [DO_SIGNUP_SUCCESS]: (state, action) => ({
-    ...state,
-    signup: asyncState.success(action.payload),
-  }),
-  [DO_SIGNUP_FAIL]: (state, action) => ({
-    ...state,
-    signup: asyncState.error(initialSingup, action.payload),
-  }),
   [SAVE_USERINFO]: (state, action) => ({
     ...state,
     userInfo: asyncState.load(initalUserInfo),
