+++ src/frontend/modules/user/types.ts
@@ -12,7 +12,6 @@ interface IUserInfo {
   role: number;
 }
 export interface IUserState {
-  signup: AsyncState<{}, Error>;
   userInfo: AsyncState<IUserInfo, Error>;
 }
